Demonstrates how many-time pad keys are easily broken
A stream cipher should not use the same key more than once. By
exclusive-or-ing two ciphertexts which use the same key, you will get
the exclusive or of the two plaintexts. You can then guess parts of one
of the plaintext and add it to the xor, thus revealing parts of the
other plaintext. This completely breaks the encryption as you can
directly find information on the plaintexts, which goes against
fundamental principles in cryptography.